cleanup history_meta to allow property access on history models
commit best viewed with whitespace diff disabled (-w)

this file was pulled from a sqlalchemy example, and most of this was
used to support things we've never implemented, such as table
inheritance and polymorphic models. We can remove most of the complexity
of this file without impacting any of our history operations since we
only have pretty simple history tables.

This also fixes an actual issue we were experiencing with history models
where you couldn't access columns on the history model that are normally
accessed via a property.

We use properties to limit/alter how we interact with db columns. This
is used in a few different ways, one common pattern is to encrypt/hash a
value in the setter. (eg _personalisation). We also use it to prevent
secret values from being read (verify codes, user passwords), and now we
use it in Service to set email_sender_local_part automatically when
custom_email_sender_name is set.

The actual column itself on the main object is of the format

`_col_name = Column("col_name", ....`

The history_meta Versioned decorator loops through the columns
recreating and populating them on the history model. Previously, the
create_history function would populate the value based on the variable
name - this meant to access the data you'd need to use
`HistoryModel._col_name`. The property itself isn't copied over.

This commit changes this behaviour, so instead, the underlying column on
the history table is now a variable with a name that matches its
database column name.

While previously, we had documented the old behaviour on the callback
api bearer tokens, this is less than ideal for the service history table
where we're more likely to want to see what the historical sender was
for a service. This change makes it easier to access these properties
on audit tables while keeping the code as simple as possible.

The new state of play is as follows

```
Model.column_name <-- property/hybrid_property
Model._column_name <-- column (data as stored in db)
ModelHistory.column_name <-- column (data as stored in db)
ModelHistory._column_name <-- AttrError this does not exist!
```