Fix POST template schema for bilingual letters
The previous schema had a flaw: if `letter_languages` was not passed,
the schema was unintentionally requiring `letter_welsh_subject` and
`letter_welsh_content` to be null values, if present.

The new schema allows any of the following options:

* The three "letter" language keys to be missing
* `letter_languages` to be missing, in which case the pair of
`letter_welsh_subject` and `letter_welsh_content` can be present,
together, with matching data types (both null or both strings).
* `letter_languages` to be `english`, then both `letter_welsh_subject`
and `letter_welsh_content` must be present and null
* `letter_languages` to be `welsh_then_english`, then both
`letter_welsh_subject` and `letter_welsh_content` must be present and
strings