migrations: add 0438_min_numeric_scl_bulk
this updates the numerics in notifications, notification_history
and ft_billing to their min_scale following the DMS migration.

for ft_billing and notifications, performs updates in blocks
(partitioned on a pseudorandom uuid) to avoid locking any whole
tables (or even one specific time-period) at once.

for notification_history, partition the blocks on created_at as
that is more closely correlated to the physical layout of the table,
and accesses to it (outside the periodic update tasks) aren't
particularly clustered around any one point in time that we should
avoid locking.