commands: add commands for updating numeric scale for large tables
these commands update the numerics in Notification,
NotificationHistory and FactBilling to their min_scale
following the DMS migration.

for FactBilling and Notification, performs updates in blocks
(partitioned on a pseudorandom uuid) to avoid locking any whole
tables (or even one specific time-period) at once.

for NotificationHistory, partition the blocks on created_at as
that is more closely correlated to the physical layout of the table,
and accesses to it (outside the periodic update tasks) aren't
particularly clustered around any one point in time that we should
avoid locking.