add not null migrations for email local part
this is a super complex and weird process. I've tried to break it into
five (!) separate migrations, partially to keep each migration easy to read
and also to keep transactions small and easy to reason about

First, create check constraints that check for not null with the
postgresql_not_valid flag[^1]. This creates a check constraint but
doesn't apply it retroactively to old rows. We do this across two
migrations (432 and 433) so that the locks are acquired on separate
tables in separate transactions to avoid the risk of deadlocks.

Then validate that constraint in a separate migration (so that migration
can run in a separate transaction while it gets low level locks to
validate the constraint).

Then alter the column to have a not null constraint on the column - this
doesn't have a "not valid" option in postgres, but from postgres 12
onwards they added a shortcut where if a constraint exists that ensures
that no null values will exist on that column, then postgres is smart
enough not to scan the table[^2]. Then once that not null is set on the
column, we can remove the check constraint that we used to help set it
up. Again, do this across migrations 435 and 436 to keep access
exclusive locks on their own.

We could just have the check constraint on its own and save a couple of
migrations, but the not null on the column is nice for two reasons:

* It means our column can have `nullable=False` on the model which helps
  as an aide memoire when writing code
* It's more efficient on the postgres side.

[^1]: https://docs.sqlalchemy.org/en/20/dialects/postgresql.html#postgresql-constraint-options
[^2]: https://www.postgresql.org/docs/16/sql-altertable.html#:~:text=if%20a%20valid%20CHECK%20constraint%20is%20found%20which%20proves%20no%20NULL%20can%20exist%2C%20then%20the%20table%20scan%20is%20skipped
[^3]: https://www.postgresql.org/docs/current/ddl-constraints.html#:~:text=A%20not%2Dnull%20constraint%20is,null%20constraint%20is%20more%20efficient.