remove access to normalised_service_name
we don't actually use this anywhere anymore - just to make sure we don't
think about it at all, remove all reference to it. We use
email_sender_local_part whenever we're thinking about actual email
senders, and we only need normalised_service_name to be kept up to date
when a user is updating a service's name so we can check that name for
uniqueness.

(As a future improvement, we could simply make `normalised_service_name`
a check constraint or computed column or similar)