Remove dependence on `notifications_utils.polygons`
So we can remove the emergency-alerts-related code from utils without
having to do the bigger job of removing emergency alerts from the API
entirely.

This means that no simplification will happen on polygons posted to our
external API (which not being used at all as far as I can know).