Remove dependence on `notifications_utils.template.BroadcastMessageTemplate`
So we can remove the emergency-alerts-related code from utils without
having to do the bigger job of removing emergency alerts from the API
entirely.

This means that emergency alert templates will be treated as text message
templates for the purposes of validating length. This does mean it would
be possible to   create a template that’s too long if using
Welsh-specific characters but this shouldn’t cause any problems unless
it goes to an MNO (in which case we have a bigger problem).