run the letter still pending virus scan check every ten minutes
currently we run every hour, and wait for a letter to be stuck for 90
minutes before we kick it off again.

We don't think this is necessary, and is causing issues where, for
example, a letter passed virus scan and was sent to template preview for
sanitising at 4:30pm. the template preview task crashed, and we didn't
pick up the letter to scan/sanitise it again until 7pm, causing it to
miss the cutoff.

it isn't the end of the world if a letter is processed/scanned twice as
this should be relatively idempotent, so we can just check every ten
minutes for safety