Let `CHECK_PROXY_HEADER` be set with an environment variable
This lets `CHECK_PROXY_HEADER` be set with an environment variable in
staging and production instead of always being set to `True`. We don't
need this value for the ECS apps and were seeing errors when
document-download-frontend tried to make authenticated requests to the
api since the header was not there.