Allow creating letters with bilingual fields
When we added support for bilingual letters, we didn't allow creating
them (via the admin API) from scratch. You would need to create a basic
letter template, then POST an update with bilingual fields.

In order to allow efficient copying of templates, let's make it possible
to create a letter template with bilingual fields in a single request.