Ensure we update the right type of service callback
We have a bug where editing the delivery status callback for a service
can actually overwrite the complaint callback instead. We don't have a
proper UI for changing or adding complaing callbacks - there is only one
team that has this set up.

The `get_service_callback_api` function wasn't filtering by callback
type, so could return the delivery status callback or could return the
complaint callback, leading to the wrong type being updated. This adds
an extra filter to the function so that we have to specify the type of
callback we are updating.