Ensure serialized services only return delivery status callbacks
The `service_callback_api` of a service can return a list containing
all the entries in the `service_callback_api` table. For most services
this is just the delivery status callback but we have one service that
also has complaints callbacks set up, and so we were returning a list
with two items for this service. Admin just picks the first item in the
list, so can sometimes get the wrong one.

Serialized services now only return the delivery status callback so that
admin doesn't accidentally use the complaint callback (which we haven't
built the UI for).