Remove SQLALCHEMY env var
This var is set within the "concourse_tests" layer of the Dockerfile.
it is used to set connection parameters for postgresql.

I have also removed it from the docker run command, where I also added the "-d" parameter to the run command to detach it and make the tests run available as output.