Adjust values in Redis if letters are cancelled
We store a Redis key for the total daily notifications sent by a service
and for each notification type so that we can ensure that services don't
exceed their rate limits. If a large batch of letters is cancelled, we
should adjust these values in Redis so that cancelled letters do not
count as part of the rate limit.

This adds a function to update the values in Redis for total daily
notifications sent and total daily letters sent by service if a letter
job is cancelled. This could be added to the endpoint to cancel a single
letter, but cancelling individual letters is unlikely to cause a service
to falsely appear to go over its limit.

It makes use of a new `decrby` method in the `RedisClient`. It would be
possible to just use `set` to re-set the value, however the benefit of
using `decrby` is that the TTL stays the same - 1 day after the key was
first stored.