Only adjust cache for letters cancelled today
We decrease the count of letters and total notifications sent stored in
Redis if letters are cancelled. However, we create a new cache key each
time the date changes - cache keys have the format of
`f"{service_id}-{yyyy_mm_dd}-count"` and
`f"{service_id}-{notification_type}-{yyyy_mm_dd}-count"`.

If someone is cancelling letters sent after 17.30 yesterday, those
totals will be in stored in yesterday's cache key so we don't want to
adjust the cache for today's date. We're no longer checking the cache
key for yesterday (if it still exists) so we don't need to take any
action.

We now only update the cache if today's letters are being cancelled.