Rewrite Dockerfile for tests with postgres
The newly created dockerfile based on the tests layer was calling MAKE BOOTSRAP.
That does a bit more than we need/want since we already install the python dependencies and we already generate the version file.
It also tries to create the notifications-api DB.
But it cannot create a db at the tests layer as postgres is not yet installed.
So basing our concourse_tests layer on it was not a good option.
We have switched to using the PRODUCTION layer and adding the same necessary components for the tests.
We eliminated the MAKE BOOTSTRAP command
We install postgres, configure username and passwd and create a db for tests
We have moved the FLASK DB UPGRADE to the pipeline to seat just before the tests.