Adding a command to copy the tests folder to the image
The tests folder was not being copied to the image.
I got mistaken by the "run_locally_with_docker,sh" script because it maps my local directory to the image. So locally i could see the tests folder and run tests.
The extra line now fixes the copying of this folder to the image