Set reply to on invite request as user requesting
When a user requests to join a service, the person responding to that
request might want to follow up with the user.

By making the reply-to address the email address of the user making the
request they can do this just by hitting reply.

We will also add content to the email later telling them they can do
this.

It also means that if the email bounces then that message goes back to
the user who made the request, not us.