send_one_off_notification: don't eager-load users on initial load of service
users aren't needed until the recipient is validated, at which
point service_allowed_to_send_to does a re-fetch of the service
anyway. so whether or not performing an eager load of all users
is the most efficient way of performing this check, there's no
need to do it twice.

this is likely what was causing some requests to fail on ECS
(though the exact mechanism is unclear) - but all services for
which this view caused a segfault had a ridiculous number of
users (>500).