fix bug with setting letter contact block deleting some fields
if you update a template to update the sms sender, email reply_to, or
letter contact_block, we call a special function
`dao_update_template_reply_to`.

this function has not been kept up to date with changes to the template
model, and as a result, when people edited their templates, other bits
of data got removed and left the template in an inconsistent state.

the reasons we added this function are complex and to do with nuances
of versioning that i haven't fully digested[^1]
this bug has appeared before[^2]

We should fix this function properly in the near future, but for now the
priority is to make sure this doesn't reoccur.

[^1]: https://github.com/alphagov/notifications-api/pull/1556
[^2]: https://github.com/alphagov/notifications-api/pull/2347

we should fix this properly by refactoring update_template