Check every field to see if a template has changed
In https://github.com/alphagov/notifications-api/pull/361/files we
introduced a check to see if a template had changed before saving a new
version of it.

This pull request introduced a list of fields to check for changes.
This list needed to be maintained and added to over time, otherwise
changes to a new field wouldn’t be considered a material change.

Rather than maintaining a custom list of which fields we should check
for changes, let’s check all of them.

Things which change a template but don’t create a new version (for
example moving it from one folder to another) go through different
routes anyway, so there isn’t a concern about accidentally creating
new versions there.