Late-assign IDs to objects with history
Objects that have a corresponding `…History` object must have a value
for their `id` property, so that the two objects can be matched up.

Previously we were manually assigning this `id` in the DAO layer.

This creates a problem when the REST layer also creates objects which
have a foreign key relationship to an object which has not yet been
persisted by the DAO layer.

What happens in this case is:
1. primary object gets created (without an ID)
2. secondary object gets created, forcing SQLAlchemy to give an ID to
   the primary object to define the foreign key relationship
3. SQLAlchemy eagerly commits
3. primary object is passed to the DAO layer, where it is given a new ID
4. DAO layer commits the object again, violating a foreign key
   constraint because the secondary object now refers to an ID which
   doesn’t exist

The way we have got around this before is by accessing unrelated
properties on the primary object, which seem to prevent step 3 from
happening. An example of this is the comment and subsequent line here:
https://github.com/alphagov/notifications-api/blob/492143312cd5ad75775f49c8fd7d3037370115fd/app/template/rest.py#L82-L83

This commit moves the assigning of `id`s into the code which creates the
history object.

This means that:
- it is done consistently in one place
- it is only done where objects don’t already have a `id`, so we don’t
  change an objects `id`