try and improve template delete performance
Background: When deleting templates, we find out when it was most
recently used so that we can inform the user.

Problem: This query can be extremely slow, sometimes timing out. A
couple of random picks showed it to be slow when templates haven't been
used recently at all (so it needs to search through lots of
notifications to see if there's a match so it can find the most recent
created_at timestamp)

Reckon: templates that people want to delete haven't been used recently.

Theory: if we search using the `ix_notifications_template_id` index just
to see if any row exists at all, we can bypass this query for a decent
chunk of template delete queries, thus improving performance.