Remove custom equality check
Since we’re checking all fields of the dictionary now, this is
equivalent to checking `dict` equality (at least in the case where the
keys are the same for both dicts, which they are because they both come
from `TemplateSchema.dump`).

So we no longer need the custom generator expression or the function.