Fix check to see if migrations should run
`API_HOST_NAME` is set to the full URL in the Concourse pipelines, so it
is already in a format like `https://api.notify.works`. By setting it to
`API_HOST_NAME=https://api.${DNS_NAME}` we were ending up with it being
set to `https://api./_status`, and the check to see if migrations should
run was always returning true.