bump boto3 (and celery0 to bring in latest ses API
we need to bump boto3 to bring in the changes to the sesv2 API to add
support for custom headers.

I went through the boto3 changelogs with a fine toothcomb - nothing
there is particularly alarming. Here's a filtered list of just the lines
i think will affect our apps. Note that we only use boto3 for
interacting with things programatically, so while there may have been
changes to, eg, boto3's ecs API, since we don't interact with ECS from
our apps we don't need to worry about that. See the filtered list[^1].

Notably there were some changes to the boto3 sqs protocol, and celery
needed to be bumped to support that - see the kombu PR[^2] if you're
interested in learning more.

[^1]: https://gist.github.com/leohemsted/0572ad15c57fdbaef8b9697cc6e5eaa4
[^2]: https://github.com/celery/kombu/pull/1807