make reply_to_address non-optional
it was both Optional (in that you can pass in a string or None), and
also optional (in that you don't need to pass anything in at all).

Doing one but not the other makes the function a little cleaner. We
only call this function from one place, and that always passes in a
value.