Add column to User table to show if user receives features email
This adds the `receives_new_features_email` column to the users table so
that we can store whether or not people want to receive emails about new
features in the database. It has a default of true, but we will need to
manually update the data before using it since we have a list of people
who have opted out.

While this migration will require an access exclusive lock, it does pass
the Squawk linting and it is running it on a small table.