Create model for unsubscribe_request table
table links to notification, service, templates_history (and templates).
it has a batch id that links to - see chris's mockups for info.

this also includes a unsubscribe_request_report. this is a mouthful to
say so i'm keen to think about better names for it, perhaps involving
"batch" as we've referred to that in UI mockups.