Add script to lint migrations from Concourse
The existing script to lint migrations using Squawk won't work when run
from Concourse. This renames it (with no changes) and creates a new
script for running Squawk on Concourse.

When running from Concourse, we can't use Git to compare the migrations
on the branch to the migrations on main. The pull-request resource we
use clones main at the latest commit and merges the pull request at
into main, so the branch no longer exists. The script for Concourse
checks the latest version from GitHub instead.