move FROM_NUMBER override from root Config to Development
We already had a manually set FROM_NUMBER of `development`, but when
we updated the root config to be environment-agnostic[^1] we missed
this. As such, when FROM_NUMBER isn't set manually in the environment
(which it isn't on dev envs), we'd error because it'd default to None
rather than to `development`.

Move this declaration to the `Development` config where it fits anyway

[^1]: https://github.com/alphagov/notifications-api/pull/4078