Make dao_get_last_date_template_was_used only search within last year
To reduce time it takes to run the query and prevent delete_template
page from timing out.

This change is paired with content change in admin - we used to
say that content was never used if we found no records, no we
say it was not used within the last year.