set NotificationAllTimeView relationships viewonly
NotificationAllTimeView is a view (quelle surprise). You can't edit a
view rather than the underlying table.

Sqlalchemy doesn't have a way to mark a model as a view, so there's
nothing stopping you writing `NotificationAllTimeView(id=..., ...)`.
This would error when sqlalchemy tried to execute the sql.

However, as sqlaclhemy doesn't know this it's printing out lots of
warnings about UnsubscribeRequest.notification conflicting with
NotificationLetterDespatch.notification. We can solve these warnings by
marking the relationships to the NotificationAllTimeView as viewonly.

For an explanation on why this is an issue more generally, see the
following commit.