Set UnsubscribeRequest.template to viewonly
Without this, sqlalchemy prints warnings about a conflict between
`UnsubscribeRequest.template` and `UnsubscribeRequest.template_history`
(and also `TemplateHistory.unsubscribe_requests`, the backref for the
second of those relationships).

The conflict occurs if you create a sqlalchemy ORM object by specifying
relationships rather than columns, and relying on sqlalchemy to set the
column values themselves. Imagine the following situation:

```python
my_template = Template(id=1234, version=1)
other_template_history = TemplateHistory(id=4567, version=2)
ur = UnsubscribeRequest(
    template=some_template,
    template_history=other_template_history
)

assert ur.template_id == my_template.id
assert ur.template_id == other_template_history.id
```

This behaviour is undefined by sqlalchemy and it'll randomly pick one of
those to set on the db object.

The solution to this is to set the `template` relationship as viewonly.
This means you cannot write `UnsubscribeRequest(template=some_template)`
at all - this is useful as the relationship between an unsubscribe
request and a template spans _all_ versions of that template ("lets see
all historical unsubscribe requests across all versions of this
template") so with that code snippet we wouldn't know what
`template_version` to set on the UnsubscribeRequest object.

When creating an UnsubscribeRequest, we should be creating it from a
notification, which has a link to a specific TemplateHistory object.