Add serialize_with_billing_info() to Notification model
This is so that we can send our users billing information along
with other information about the notification when they
send us GET notification request.

We return following information:
cost_in_pounds: overall cost of the notification, sent as a string
to avoid floating point issues. Alternatively we could return
a dictionary with pounds and pence as separate key-value pairs.

I decided to go with a string, as I saw paypal doing the same.
I think either would be ok in the end. String is more human-readable I think,
as this is how we would see prices in receipts, and it should be easy
 to process across different languages.

cost_details: extra info about the cost, which is:

for letters: sheets_of_paper, postage
for sms: sms_fragments, rate_multiplier, rate
for emails: empty dict
For cost_details, we are getting rate information from other tables - rates
and letter_rates tables.

I followed a pattern we have on User model, where we check permissions
table in User.get_permissions().

These rates tables are very small, so hopefully checking them shouldn't
cost us too much, but if you think it's a problem, I'm happy to think of
alternative ways - maybe connect those two tables with notifications table,
so they are readily available?

For letters in "pending-virus-check" status, that information won't be ready yet.
So there is a boolean "is_cost_data_ready" field to indicate if
cost information is ready. If it's not, we return None for cost,
and empty dict for cost details.