Account for billing data not being ready for freshly created templated letters.
When we create a templated letter, we first persist the notification,
and only later, asynchronously, we create a PDF for that letter.
we know how many pages the letter is once the PDF is created for it.

So for a little bit, we have a letter in created state without knowing
how much it will cost. The code has now been changed to reflect that.

I also added a parametrized test to illustrate how letter cost info
behaves for letters that were not sent - for example cancelled letters,
letters that failed validation, or letters that failed due to some
other issues.

Last but not least I renamed sheets_of_paper to billable_sheets_of_paper.
It's a bit long, but less confusing in those cases where the letter
doesn't get sent.