In a test, represent rate_multiplier as a float not int
This is because in the db we store rate_multiplier as db.Numeric
(where for example billable_units are stored as an db.Integer).

I assume this is so we can have fractional rate multipliers if
need arises?

That's why I opted for reflecting how we store the data in
what we send to the users, and how it's represented in the test,
instead of casting it as int when serializing for the users.