Revert "Log service ID as part of sending callbacks if possible"
This reverts commit b877f233945a2fc4680f3f3fed4f8520eee26b0f.

The service ID isn't being logged. Although the service id is
being added to `extra`, it isn't appearing in our logs. My hunch
is that the it is being replaced afterwards by our generic
service-id adding logging:
https://github.com/alphagov/notifications-utils/blob/main/notifications_utils/logging/__init__.py#L217

For the moment, let's revert this code because it isn't doing
anything and then I'll hope to find time to find a fix/alternative
for this.