Pin requests due to bug with connecting to DVLA
We updated requests in https://github.com/alphagov/notifications-api/pull/4108
but we saw errors in production connecting to DVLA as per
https://github.com/psf/requests/issues/6730.

It looks like moving back to 2.32.2 should be unaffected, at least
according to the github issue above.

Once a new version is released that fixes it, we can drop this
pin. If it doesn't get fixed we could look at an alternative approach
like https://github.com/alphagov/notifications-api/pull/4113