Remove org permissions when archiving a user
When archiving a user, we were removing them from their services and
removing all service permissions. We were also removing them from any
organisations that they were a member of, but we were not removing their
organisation permissions which gave an error when trying to serialize
the user.

This change ensures we now remove an archived user's organisation
permissions too.