Cost is zero for letters that were never sent
Even if they have billable units set.

I have included technical-failure in this, but mot permanent-failure,
because I only remember us setting letters to permanent-failure
when not sending them was due to user error, and they were still
charged for these letters as far as I remember.

Also fix one of the tests that were broken following introducting
check if cost data is ready for sms.