When getting rate from redis cache, cast it to float
Even though we send the rate as a float to Redis, redis then
gives it back to us as a byte-string.

We did not take it into consideration when building the test data,
and didn't have redis server running when testing on local,
and it broke our deploy sadly.

So now we update the unit tests mocks to behave more realistically,
and we cast the rates we are getting back from redis to float to
fix the issue.