Fix postage validation in schemas
The postage validator allowed the value to be "first", "second",
"europe" or "rest-of-world", however when sending a letter using the API
or setting the postage on a template you should not be able to use the
international postage options.

We use the postage validator in a number of places and there is one of
those that does need to allow international postage - when the admin app
sends a precompiled letter it can use the international postage types.
This splits the postage validator into two - one for the UK types of
postage (used in most places) and one for the admin app to use (used in
one place).