Count all unsubscribe requests from last 7 days on dashboard
The dashboard says:
> In the last 7 days
>
> 1,234 email unsubscribe requests
> Latest report 3 weeks ago

At the moment the 1,234 number will go down when requests are marked as
completed. This doesn’t feel right, because those requests have still
been received in the last 7 days.

This commit changes the query to count all requests from the last 7
days, whether or not they have been marked as completed.

This is consistent with received text messages and returned letters
(although we don’t have a similar concept of marking these are
completed at the moment).