Send callback URL to DVLA when sending letters
We want to pass in a callback URL when making the API request to send a
letter to the DVLA. This setting is controlled by an environment
variable, since the callback funtionality isn't ready yet. Each callback
URL contains the signed notification ID in the query string so that we
can check the callbacks we get sent are legitimate.