Generate unsubscribe link at time of sending
We need to store the unsubscribe link where it is something that a user
has provided to us through an API call.

But where we are generating the link we don’t need to store it, because
we already store all the information needed to generate it.

By not storing it we make it easy to differentiate which notifications
have user-supplied unsubscribe links and which have ones which we are
generating.