Add unsubscribe link to body of emails
We have started sending list-unsubscribe headers if a user marks a
template as unsubscribeable, for teams who aren’t using the API.

However we are already telling these teams to add unsubscribe links to
the body of their emails.

This means:

- there are two different ways to add unsubscribe links
- users will have to manage requests to unsubscribe coming from two
  different places

This will be confusing, and make it harder for us to write good guidance
that users will be able to follow, and harder to explain what the ‘let
users unsubscribe from these emails’ checkbox does.

If we have a confusing product proposition and users cant follow our
guidance then we risk our spam score rising.

This commit adds a way to generate a separate unsubscribe link which we
can add to the footer of every email (using the argument added in
https://github.com/alphagov/notifications-utils/pull/1136/files)

We can then report unsubscribes back to the team in the same way, no
matter if the recipient has clicked the button in their email client
or the link at the end of the email.

If the user has checked the checkbox then we will always show the link
in the body of the email (this feels like the most predictable
behaviour). If they have also passed in a one-click URL for the headers
then we will respect that too.

The only combination which isn’t possible is using Notify’s generated
URL for one-click, and your own in the body of the email. This feels
like the most unlikely combination.