Purge additional Redis cache when user unsubscribes
When a user unsubscribes we delete the cahce of statistics which
populates the dashboard.

We also cache the summary of unsubscribe reports that the person running
the service has batched up.

When we get a new request to unsubscribe it goes into the ‘unbatched’
report, so we need to delete the cache of reports for it to appear.