Tweak start and end of unsubscribe request reports
Originally we thought it would be nice and clean to extend the end dates
of the reports so there were no gaps between reports.

On second thought I think this will be confusing where the latest report
claims to run up to the current time. In this case:
- the report, when downloaded might not even contain requests from today
  so it could look like requests are missing from the report
- the report might claim to have requests up until now, but if any new
  ones come in between the user clicking on the page and generating the
  report then they wouldn’t be included

In could also be confusing when a report is cleaned up by data
retention. Adjacent reports would then appear to expand to fill the gap,
even though they gain no new requests.