Add log line for letters that failed validation due to no fixed abode
This adds a log line for precompiled API letters that were sent to addresses
which were no fixed abode addresses. This will allow us to look up the
address of the notification and check for false positives.

For templated letters sent via the API, we can already see from the logs if
any failed validation due to having a no fixed abode address. We won't
be able to see the address that was used, but can check which services are
sending them.